refactor: . For files now in src/Task/, move their tests to tests/Task/
Urgency.test.ts was in inconsistently tests/Query/
