+++ tests/Task/TaskLocation.test.ts
@@ -1,4 +1,4 @@
-import { TaskLocation } from '../src/Task/TaskLocation';
+import { TaskLocation } from '../../src/Task/TaskLocation';
 
 describe('TaskLocation', () => {
     it('should store the full task location', () => {
